rename sim to sitl module
