+++ payload_controller/main/CMakeLists.txt
@@ -2,7 +2,7 @@
 SET (priv_requires imu encoder)
 
 if(IDF_TARGET STREQUAL "linux")
-    list(APPEND PRIV_REQUIRES_COMPONENTS "sim")
+    list(APPEND PRIV_REQUIRES_COMPONENTS "sitl")
 endif()
 
 idf_component_register(SRCS "payload_controller.cpp"
