+++ payload_controller/src/sitl/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(srcs sim_runtime.cpp)
+set(srcs sitl_runtime.cpp)
 
 if(IDF_TARGET STREQUAL "linux")
     idf_component_register(
