+++ payload_controller/src/sitl/sitl_runtime.cpp
@@ -1,7 +1,9 @@
-#include "sim_runtime.hpp"
+#include "sitl_runtime.hpp"
 #include "esp_log.h"
 
-namespace sim {
+static const char* TAG = "SITL_RUNTIME";
+
+namespace sitl {
     namespace {
         SimConfig config;
         bool loaded = false;
@@ -11,16 +13,25 @@ namespace sim {
 
     const SimConfig& get_config() {
         std::lock_guard<std::mutex> lock(mtx);
+
         if (loaded) {
             return config;
         }
 
         const char* gz_model = std::getenv("GZ_MODEL");
         if (gz_model == nullptr) {
-            ESP_LOGW("SIM_ENV", "No value specified for GZ_MODEL, defaulting to %s", config.gz_model.c_str());
+            ESP_LOGW(TAG, "No value specified for GZ_MODEL, defaulting to %s", config.gz_model);
         } else {
             config.gz_model = gz_model;
         }
+
+        const char* gz_world = std::getenv("GZ_WORLD");
+        if (gz_world == nullptr) {
+            ESP_LOGW(TAG, "No value specified for GZ_WORLD, defaulting to %s", config.gz_world);
+        } else {
+            config.gz_world = gz_world;
+        }
+
 
         loaded = true;;
         return config;
